Moved the SSID structs and traits in a dedicated `ssid` module
